move sidebar back in util
